Fix send_pictures extension
